fix: move test directory
